sched/cass: Avoid using the prime CPU on systems that have one
On arm64 systems with a prime CPU, treating this CPU – which is the single
fastest one in the system – the same as any other CPU is extremely bad for
energy efficiency. This is because prime CPUs are designed to be very fast
at the expense of energy efficiency, serving as the single-core performance
workhorse of a system.

Since CASS hasn't been giving special treatment to prime CPUs, CASS has
been balancing relative load onto prime CPUs at great expense to energy.

Thanks to the checks in place for CPU overload and task fit, it's easy to
adjust CASS to avoid using prime CPUs without hurting single-core or
multi-core performance.

Place a check just below the task fit check for whether or not a candidate
is a prime CPU. This way, when CPUs are overloaded or a task only fits on
the prime CPU, CASS will still utilize the prime CPU without a loss of
performance.

This provides a double-digit percent improvement to energy efficiency
across the board as measured on Tensor G3 (Pixel 8) and Tensor G4 (Pixel 9
Pro).

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: kvsnr113 <kvsnrprojkt113@gmail.com>
Signed-off-by: Gaurav Koli <79272711+koligaurav462@users.noreply.github.com>